requirements.txt
fastapi==0.115.6
uvicorn==0.32.1
pydantic==2.10.3
python-dotenv==1.0.1
requests==2.32.3

# RAG (local, simple)
sentence-transformers==3.3.1
faiss-cpu==1.13.2
numpy==2.1.3
